Fixed logging in as different user types.
When launching for the first time, the user is taken to a page where they select if they are logging in as a student or a driver. This is then passed to the settings and accessed when logging in. (Currently selecting driver takes the user directly to the driver homepage)